requirements.txt
matplotlib
numpy
pandas
torch
scikit-learn
google-colab
